Ensure we use the forwarded host header when it exists
With Fakamai integration, we receive a header: `x-rh-insights-forwarded-host`
which contains the actual hostname we need to set for our SP config, to match
the response from the IdP. In the case of CI for instance, this would be:
ci.internal.cloud.redhat.com

We need to ensure we set this for the `$proxy_host` value in order for our SP
to behave correctly, and for our redirect on `401` to use the correct hostname.

Locally, when this header is _not_ set, we should default this to use the actual
host.