Make `/api/ping-service/ping` a more useful `/api/turnpike/identity` endpoint
Ahead of having some chroming in front of Turnpike, this sets up a way for the
front-end to grab identifying information from the SSO SAML assertion.

This will return the `x-rh-identity` header as JSON if it exists, otherwise returning
a JSON payload with an `error` message if it doesn't exist, or fails to parse.

- remove `/api/ping-service/ping` mocked service
- add `/api/turnpike/identity` utility endpoint
- updates docs to reflect this endpoint for consistency
- updates the backend config for this endpoint (will need to update config in app-interface)