Ensure `x-rh-identity` header `Role` attribute is a list
Since our SAML assertion sends `Role` as a multi-value attribute like:
```
{
  "Role": ["Role1", "Role2"],
  "givenName": ["Jane"]
}
```
Our identity header would provide you with:
```
{
  "Role": ["Role1", "Role2"],
  "givenName": "Jane"
}
```
This will preserve the role list, and unpack other values with one element.

There is a case where `Role` may contain just one role, like:
```
{
  "Role": ["Role1"]
}
```

In this case, the current logic would unpack that as well, and we'd end up with
a header like:
```
{
  "Role": "Role1",
  "givenName": "Jane"
}
```

This chance will maintain a list for `Role` regardless of the number of elements.
Since this was a customization on the SAML assertion from the IAM team, we should
be able to rely on other fields always being single-value or multi-value, though
we may want to confirm this.

Other minor changes:
- comments for CI deployment
- pre-populate `.env.example` with `FLASK_ENV=development` for debugging