Fix NGINX startup failures
When the NGINX server starts, and one or more upstream hosts can't be resolved,
NGINX fails to come online at all. This means pods will enter a crash loop, and
Turnpike will be down in the given environment.

To fix the issue, we need to ensure NGINX doesn't try to resolve all hosts at start,
and setting the `proxy_pass` value to a variable will lazily resolve the host.

The following changes work locally, and have been deployed and tested in an ephemeral
environment (outside the context of a route/Akamai, just confirming NGINX starts).

- Configure a local regression API which will cause NGINX to fail to start if it
  doesn't resolve without this fix in place.
- Update the location matcher so that we can explicitly proxy the entire request
  path, which doesn't happen when you use a variable in the config.
- Use the internal Docker DNS resolver IP for the NGINX resolver.
- Set the `$upstream` to proxy to in a variable so that resolution doesn't happen
  at startup. This will allow for other services to be online, while the failing
  service(s) will throw a 5xx until the resolver picks up the changes, assuming
  it comes back online.