feature: improve back end matching
We were facing a problem which some back ends were not getting templated
as Nginx locations because they had some sort of invalid or unexpected
configuration. On top of that, we were seing that the back ends matched
by Nginx and the ones matched by Turnpike were different.

In order to avoid the issue with the missing locations, the script that
builds the configuration for Nginx now raises errors and stops the
execution whenever there are back ends that are not properly defined.
That will help us detect misconfigurations.

For the back end mismatching issue, I realized that we were using regex
locations for Nginx which were working fine for the vast majority of the
locations, but not so well for back ends that had similar routes. For
example, the following routes...

- /api/v1
- /api/v1/service
- /api/v1/service/path

... would make Nginx return on the first match even for a request for
the "/api/v1/service/path", because that's how regex based locations
work in Nginx. Since we are not using any fancy regex in our locations,
I switched back to using no modifiers for them, which will ensure that
Nginx keeps looking for the longest match before returning.

RHCLOUD-33532