fix: the "oidc" plugin is not skipped
When a back end did not have an "oidc" section in the "auth" section,
the plug in was returning "500" errors, when it should simply be
skipping the authorization attempt for that back end.

The reason for this was that I mistakenly understood that if a plug in
gets executed, that meant that the back end had an authorization section
defined for that plug in.

However, in reality, Turnpike loops through the plug ins and skips the
authorization when there is no section defined in the "auth" key for
that plugin.

RHCLOUD-30610