feature: remove dependency with the Flask application to build config
The Flask application provides a few headers and a route which are
dynamically templated into the Nginx configuration whenever the pod
launches. However these configurations have been static for a long time,
and for newcomers and maintainers having that dynamic configuration just
adds confusion.

These changes hard code the headers and the routes in the Nginx
configuration which also allows removing the pieces in the code that
provided them from the Flask application.

The Nginx application no longer will have to wait for the Flask
application to boot.

The changes also clean up and document the current Nginx configuration.

RHCLOUD-40755