fix: partially revert Nginx location matching
The reason why regex was being used for the location matching was
because the captured group was being passed to the "proxy_pass" as the
"$1" argument. So, for a backend like this one...

- Origin: http://my-service.svc.cluster.local
- Route: /api/my-service

... and a request sent to "/api/my-service/resource/subresource", the
following URL was constructed in Nginx:

- http://my-service.svc.cluster.local/resource/subresource

I still believe it would be benefitial to move away from regex matchings
since it would make our lives easier, but until we figure out a solution
to appending the URI I'll revert the changes to unblock other teams.

RHCLOUD-33532