feature: rewrite URI for the proxy pass directive
The only reason we were using regex locations for Nginx was to be able
to strip the route suffix from the incoming request's path, and append
the rest of the route to the defined upstream's origin. For a back end
defined like so...

- Origin: http://my-service.svc.local:8080/endpoint/
- Route: /api/my-service/endpoint

... and a request sent to
"/api/my-service/endpoint/resource/subresource", we were stripping the
prefix and appending the rest of the path, "/resource/subresource", to
the origin.This resulted in Nginx sending a request to the following
route after a successful authentication:
"http://my-service.svc.local:8080/endpoint/resource/subresource".

However, having regex locations for Nginx was causing issues because for
overlapping URLs, Nginx simply returns the first match. So, with the
following locations defined...

- location ~ /api/base
- location ~ /api/base/path-a/
- location ~ /api/base/path-a/subpath-a

... and a request sent to "/api/base/path-a/subpath-a/subsubpath-a",
Nginx would match the first regex location and return, instead of
attempting to match the longest defined location. This is the reason why
we are moving to default Nginx locations, which follow a "longest
match" approach, and using a "rewrite" directive instead which will
allow us to perform the route stripping described above.

RHCLOUD-33532