feature: make Nginx redirect to the internal VPN host
For requests that were coming from the VPN, Nginx was redirecting login
requests to the internal host, instead of keeping the original private
host intact.

These changes modify the VPN plug in so that the Flask application has a
way of signaling the Nginx server which host should the requester be
redirected to.

It also removes the "FLASK_SERVICE_NAME" from the Nginx deployment
because it is not really needed and it adds confusion to the
configuration. With that environment variable it seemed like we were
redirecting the requester to the Flask application, but in reality we
were just redirecting them to the Nginx proxy again, but to a different
URL.

I also added another minor change that fixes the "X-Forwarded-Host"
header, which for some reason it was being overriden to the Nginx's
host, instead of keeping the original host that comes from the request.

RHCLOUD-37506