fix: all requests with no Authorization header get rejected
When a back end had an OIDC service account specified, but the requests
were being authorized by other means, the requests were getting rejected
all the time because they did not come with the "Authorization" header.

In the case that a request uses a different method than OIDC for a back
end, the plug in needs to return the context so other authentication
plug ins can be used instead.

RHCLOUD-30610