feature: set up a SAML client for the VPN authentications
Unfortunately we cannot reuse the existing SAML client to be able to
trigger the SAML SSO login flow correctly, since the currently existing
client is configured for a specific hostname, and we cannot include more
hostnames in the service provider's URLs.

These changes store a flag in Flask's session to pull the correct SAML
configuration when authenticating requests.

RHCLOUD-37506