fix: rewritten URI has a zero length
When the capturing group ends up being an empty string, Nginx complains
by raising an error that says "the rewritten URI has a zero length". For
example, for a location defined like this...

- location /api/test

... and a request sent to "/api/test", the capturing group ends up
capturing nothing, and therefore there is nowhere Nginx can rewrite the
request to.

By prepending the "$upstream" variable we ensure that for these cases
the origin defined in the back end is at least present in the rewrite.

RHCLOUD-33532