feature: make Turnpike handle "internal" and "private" SAML auths
In order to be able to provide SAML authentication for both the
"internal" and "private" networks, the Flask application required two
SAML clients set up.

We attempted using Flask's session to build the proper request context
depending on whether the user was sending the request from the
"internal" or the "private" network. However, as soon as the user
attempted logging in by hitting the SAML endpoint from the "private"
network, we set the session flag in Flask's session and the subsequent
requests —regardless of which network the request came from— would be
redirected to the "private" network, because the "OneLogin" utility
would generate that redirect from the "private" settings.

These changes split the SAML views and the plug in in different files to
make it easier to reason about how the whole thing works, and to
decouple the plug in from the views themselves — because even though
they're somewhat related, the views and the plug in are hit in different
stages of the authentication. The views are accessed when the user has
to log in, and the plug in just verifies that the user has been
authenticated.

The views are also refactored so that they get can accept the SAML
settings' type they are going to work with, in order to build the proper
SAML context accordingly, and thus removing the need to use Flask's
session for it.

It also simplifies the Nginx configuration by removing the "login URL"
header, because it is really not needed. The only plug in that redirects
the user to a login page is, and has always been, the SAML plug in. And
we can achieve the same redirect by using Nginx's built in directives
and variables.

RHCLOUD-37506
RHCLOUD-41155