fix: alternative auth schemes are not skipped
The OIDC plug in was making every request with an "Authorization" header
but with a different authentication scheme, like "Basic", to be
rejected with a "Forbidden" response.

However, the OIDC plug in should only care about "Bearer" authentication
schemes, and should let the rest of the plug ins attempt authorizing the
request.

RHCLOUD-30610