feature: make VPN plugin always set header flag for Nginx
Nginx uses a header flag to determine to which URL the users should be
redirected when logging in. The VPN plugin only set that header flag if
the back end had a "private: true" directive, but we actually require
the plugin to set it every time a request comes from the internal
network.

RHCLOUD-37506