fix: MTLS requests denied for non-private backends
I did not know that Akamai sets the "X-Rh-Edge-Host" header for MTLS
authentications too, and due to that mistake some of the requests are
being rejected.

This fix should still add the Nginx header flag for all the requests
that come from the VPN, and let the rest of the requests through.

RHCLOUD-37506